refactor(mon-pix): migrate unit tests for PATH /api/admin/users/{id}
